% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimf.R
\name{autoplot.ssimf}
\alias{autoplot.ssimf}
\title{Plot object class ssimf}
\usage{
\method{autoplot}{ssimf}(x, type = c("decomposition", "test"))
}
\arguments{
\item{x}{object with class attribute 'ssimf'}

\item{type}{character string giving the type of plot to be computed. 
Allowed values are "decompositon" (the default), "test" for signficance test}
}
\value{
plot significance graph. Shaded area isn't significance.
}
\description{
Plot object class ssimf
}

