% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimf.R
\name{criticalvalue}
\alias{criticalvalue}
\title{Critical Value}
\usage{
criticalvalue(imfs, percenta)
}
\arguments{
\item{imfs}{the true IMFs from running EMD code. The first IMF must
be included. The trend is not included.}

\item{percenta}{a parameter having a value between 0.0 ~ 1.0, e.g., 0.05 
represents 95#' confidence level (upper bound); and 0.95 
represents 5#' confidence level (lower bound)}
}
\value{
sigline: a two column matrix, with the first column the natural
logarithm of mean period, and the second column the
natural logarithm of mean energy for significance line
}
\description{
that is used to obtain the "percenta" line based on Wu and
Huang (2004).
}
\details{
For this program to work well, the minimum data length is 36
}
\references{
Wu, Z., and N. E Huang (2008). Ensemble Empirical Mode Decomposition: a noise-assisted data analysis method.
Advances in Adaptive Data Analysis. Vol.1, No.1. 1-41.
}

