% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimf.R
\name{dist_value}
\alias{dist_value}
\title{Distance Value}
\usage{
dist_value(yPos, yBar, nDof)
}
\arguments{
\item{yPos}{An input array at which PDF values are calculated---y value
yPos is an 1D 5000 pt matrix for y value in interval - [yUpper,yLower]}

\item{yBar}{The mean value of y ---exp(yBar)=E-bar}

\item{nDof}{The number of degree of freedom---Ndof=Npt}
}
\value{
PDF: a normalized output array -about the PDF distribution
}
\description{
calculate the PDF value  for 'sigline' in function critical value
}
\details{
This is a utility program being called by "confidence".
this code calculate PDF value under yPos range
within the mean value of y(y-bar) of a chi-square distribution
here main job is calculating equation(3.4)--PDF formula from the reference paper
}
\references{
Wu, Z., and N. E Huang (2008). Ensemble Empirical Mode Decomposition: a noise-assisted data analysis method.
Advances in Adaptive Data Analysis. Vol.1, No.1. 1-41.
}

