% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimf.R
\name{extrema}
\alias{extrema}
\title{Extrema
 
This is a utility program for cubic spline envelope,
the code is to  find out max values and max positions
min values and min positions}
\usage{
extrema(in_data)
}
\arguments{
\item{in_data}{Inputted data, a time series to be sifted}
}
\value{
\item{spmax}{The locations (col 1) of the maxima and its corresponding
values (col 2)}
\item{spmin}{The locations (col 1) of the minima and its corresponding 
values (col 2)}
}
\description{
Extrema
 
This is a utility program for cubic spline envelope,
the code is to  find out max values and max positions
min values and min positions
}
\details{
EMD uses Cubic Spline to be the Maximun and Minimum Envelope for
the data.Besides finding spline,end points should be noticed.
}
\references{
Wu, Z., and N. E Huang (2008). Ensemble Empirical Mode Decomposition: a noise-assisted data analysis method.
Advances in Adaptive Data Analysis. Vol.1, No.1. 1-41.
}

