% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimf.R
\name{ifndq}
\alias{ifndq}
\title{Instantanesous Frequency}
\usage{
ifndq(vimf, dt)
}
\arguments{
\item{vimf}{an IMF}

\item{dt}{time interval of the imputted data}
}
\value{
omega: instantanesous frequency, which is 2*PI/T, where T
is the period of an ascillation
}
\description{
this is a function to calculate instantaneous based on EMD method
}
\details{
this is a function to calculate instantaneous based on EMD method--
normalize the absolute values and find maximum envelope for 5 times
then calculate the Quadrature ,Phase angle,then take difference to them
finally,the instantaneous frequency values of an IMF is found.
}
\references{
Wu, Z., and N. E Huang (2008). Ensemble Empirical Mode Decomposition: a noise-assisted data analysis method.
Advances in Adaptive Data Analysis. Vol.1, No.1. 1-41.
}

