% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimf.R
\name{significance}
\alias{significance}
\title{statistic significance test for imf}
\usage{
significance(demd, alpha = 0.05)
}
\arguments{
\item{demd}{Time series object of class "mts" where 
series corresponds to IMFs of the input signal, with 
the last series being the final residual}

\item{alpha}{significance level, default 0.05}
}
\value{
A list with class attribute 'ssimf' holding the following elements:
}
\description{
statistic significance test for imf
}

