% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimf.R
\name{statistic}
\alias{statistic}
\title{Statistic}
\usage{
statistic(imfs)
}
\arguments{
\item{imfs}{the true IMFs from running EMD code. The first IMF must
be included for it is used to obtain the relative mean
energy for other IMFs. The trend is not included.}
}
\value{
\item{logep}{a two colum matrix, with the first column the natural
                logarithm of mean period, and the second column the
                natural logarithm of mean energy for all IMFs}
}
\description{
that is used to obtain the "percenta" line based on Wu and Huang (2004).
}
\details{
For this program to work well, the minimum data length is 36
}
\references{
Wu, Z., and N. E Huang (2008). Ensemble Empirical Mode Decomposition: a noise-assisted data analysis method.
Advances in Adaptive Data Analysis. Vol.1, No.1. 1-41.
}

